Crate seqalign

Structs

Matrix
A wrapper for a 2D vector.

Functions

PyInit_seqalign⚠
This autogenerated function is called by the python interpreter when importing the module.

compare_samples
Just a wrapper for the needleman_wunsch function to be used in python.

needleman_wunsch
Performs global sequence alignment of two &str using the Needleman-Wunsch classic algorithm. The returning value is a ratio whose value is the result of dividing the matches between the two aligned sequences by the length of the aligned sequences.

par_compare
Just a wrapper for the single compare and parallel_compare functions to be used in python.