Struct seqalign::Matrix

pub struct Matrix<T> { /* fields omitted */ }

A wrapper for a 2D vector.

The Matrix struct accepts any type that implements the Default trait.

Examples

    let rows = 10;
    let cols = 10;
    let mut m : Matrix<isize> = Matrix::new(rows, cols);
    m[(0, 0)] = 3;
    m[(0, 1)] = 5;
 
    let num = m[(0, 0)];
 
    let score = m.get(0, 1).expect("Error getting an immutable reference");
 
    // *score = 34; Uncommenting this line will result as an error    
 
    let score = m.get_mut(0, 1).expect("Error getting a mutable reference");
 
    *score = 10; // now m[(0, 1)] is equal to 10

Implementations

impl<T> Matrix<T> where
T: Default,

pub fn new(n: usize, m: usize) -> Matrix<T>

Creates a new instance of Matrix filling it with the default value of the type T

Example

    let mut m : Matrix<isize> = Matrix::new(10, 20);
    // Now the matrix is filled entirely with the default value
    // of the isize primitive type which is 0.

pub fn get(&self, i: usize, j: usize) -> Option<&T>

pub fn get_mut(&mut self, i: usize, j: usize) -> Option<&mut T>